Refactor CLI utilities to glyphs
Renamed and relocated CLI utility files from 'utils' to 'glyphs' directory. Adjusted corresponding imports and command names for better structure and clarity within the project.
