+++ foundrytools_cli_2/cli/glyphs/__init__.py
@@ -1,0 +1,3 @@
+from foundrytools_cli_2.cli.glyphs.cli import cli
+
+__all__ = ["cli"]
