--- foundrytools_cli_2/cli/utils/__init__.py
@@ -1,3 +1,0 @@
-from foundrytools_cli_2.cli.utils.cli import cli
-
-__all__ = ["cli"]
